+++ CHANGES.rst
@@ -4,8 +4,8 @@ Change Log
 prawcore follows `semantic versioning <http://semver.org/>`_ with the exception that
 deprecations will not be announced by a minor release.
 
-Unreleased
-----------
+2.0.0 (2021-02-23)
+------------------
 
 **Added**